Feature/main content text (#50)
* Upgrade material ui

* Create MainText component with boilerplate

* Add basic info sections

* Center credit list button

I'm not sure if styles are best declared in the credit list component or
the main text component.

* Add instruction text

* Spell bottom correctly

* Remove useless constructor

* Use AppBar for title

I feel like an `AppBar` title looks a bit tidier. I also imagine that it
could have quick buttons to the sections about the project or what to do
with it, or possibly social media links.

* Change primary color

I felt like a light red seemed gnome like.

* Punctuation

* Add shallow render test for main text

I would add a snapshot, but being so text heavy, I figure it would
change often enough to be annoying.
Since it's pretty logicless, a simple render test is good enough.